refactor: move booking MCP tools to presentation layer
book_stay, cancel_stay and list_stays move to
booking/presentation/mcp_tool/, taking the use case directly instead of the
DI container, mirroring how controllers already receive their use case.
PropertyDi/StayDi gain make<X>Tool() factories, same shape as their
make<X>Controller() counterparts.
