+++ src/booking/infra/di/property_di.ts
@@ -9,6 +9,7 @@ import type { TenantRepository } from "../../domain/repository/tenant_repository
 import { BookStayController } from "../../presentation/controller/property/book_stay.controller";
 import { CreateExternalBookingSourceController } from "../../presentation/controller/property/create_external_booking.controller";
 import { ReconcileExternalBookingController } from "../../presentation/controller/property/reconcile_external_booking.controller";
+import { makeBookStayTool } from "../../presentation/mcp_tool/book_stay.mcp_tool";
 import { ICalendarAdapter } from "../adapter/i_calendar_adapter";
 import { PostgresBookingPolicy } from "../database/postgres_policies/postgres_booking_policy";
 import { ExternalBookingSourcePostgresRepository } from "../database/postgres_repository/external_booking_source_postgres_repository";
@@ -87,5 +88,10 @@ export class PropertyDi {
     return new CreateExternalBookingSourceController(
       this.makeCreateExternalBookingSourceUseCase()
     );
+  }
+
+  // MCP Tools
+  makeBookStayTool() {
+    return makeBookStayTool(this.makeBookStayUseCase());
   }
 }
