+++ src/booking/infra/di/stay_di.ts
@@ -14,6 +14,8 @@ import { FindPropertyStaysController } from "../../presentation/controller/stay/
 import { CancelStayController } from "../../presentation/controller/stay/cancel_stay.controller";
 import { UpdateStayController } from "../../presentation/controller/stay/update_stay.controller";
 import { GetDashboardOverviewController } from "../../presentation/controller/dashboard/get_dashboard_overview.controller";
+import { makeCancelStayTool } from "../../presentation/mcp_tool/cancel_stay.mcp_tool";
+import { makeListStaysTool } from "../../presentation/mcp_tool/list_stays.mcp_tool";
 import { StayPostgresRepository } from "../database/postgres_repository/stay_postgres_repository";
 import type { TenantRepository } from "../../domain/repository/tenant_repository";
 import { TenantPostgresRepository } from "../database/postgres_repository/tenant_postgres_repository";
@@ -144,5 +146,13 @@ export class StayDi {
       this.#logger,
       this.#deviceManagementService
     );
+  }
+
+  // MCP Tools
+  makeCancelStayTool() {
+    return makeCancelStayTool(this.makeCancelStayUseCase());
+  }
+  makeListStaysTool() {
+    return makeListStaysTool(this.makeFindPropertyStaysUseCase());
   }
 }
